attempting to find error in step 2

+++ day1/test/Tests.fs
@@ -5,11 +5,11 @@ open Xunit
 open AdventOfCode2017
 
 [<Theory>]
-[<InlineData("1122",3)>]
-[<InlineData("1111",4)>]
-[<InlineData("1234",0)>]
-[<InlineData("91212129",9)>]
-let ``Solution 1 test`` (input:seq<char>, output:int) =
+[<InlineData("1122",3L)>]
+[<InlineData("1111",4L)>]
+[<InlineData("1234",0L)>]
+[<InlineData("91212129",9L)>]
+let ``Solution 1 test`` (input:seq<char>, output:int64) =
     let actual =
         let ints = input |> Seq.map (Char.GetNumericValue >> int) 
         Seq.head ints |> Seq.singleton |> Seq.append ints         
@@ -34,15 +34,15 @@ let ``Halfwise b`` () =
 
 [<Fact>]
 let ``CalcCaptcha 2 a`` () =
-    Assert.Equal(6, [(1,1);(2,2);(1,1);(2,2)] |> List.toSeq |> CalcCaptcha)
+    Assert.Equal(6L, [(1,1);(2,2);(1,1);(2,2)] |> List.toSeq |> CalcCaptcha)
 
 [<Theory>]
-[<InlineData("1212",6)>]
-[<InlineData("1221",0)>]
-[<InlineData("123425",4)>]
-[<InlineData("123123",12)>]
-[<InlineData("12131415",4)>]
-let ``Solution 2 test`` (input:seq<char>, output:int) =
+[<InlineData("1212",6L)>]
+[<InlineData("1221",0L)>]
+[<InlineData("123425",4L)>]
+[<InlineData("123123",12L)>]
+[<InlineData("12131415",4L)>]
+let ``Solution 2 test`` (input:seq<char>, output:int64) =
     let actual =
         input 
         |> Seq.map (Char.GetNumericValue >> int)
